Eixa-main/requirements.txt
# == Core do Cloud Run e Flask ==
functions-framework==3.*
Flask==2.3.3
Flask-Cors==4.0.0
httpx==0.27.0

# == Firebase e Firestore ==
firebase-admin==6.2.0
google-cloud-firestore>=2.10.0
google-cloud-storage>=2.10.0

# == Geração de Conteúdo e IA (Vertex AI/Gemini) ==
vertexai==1.49.0
google-cloud-translate==3.11.0

# == Processamento de Dados Multimídia e Texto ==
PyMuPDF==1.23.25
python-docx==1.1.0
Pillow==10.1.0
PyYAML==6.0.1

# == Utilidades e Processamento de Linguagem Natural (NLP) ==
dateparser==1.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
pytz>=2023.3

# == Sanitização ==
html5lib==1.1
bleach==6.1.0

# == Dependências Cloud Run ==
grpcio>=1.60.0
protobuf>=4.25.3
setuptools>=65.5.0